refactor(production): 抽出门身份模块，解掉 driverOps 破 800 行
修洞班的改动把 productionRunDriverOps.ts 顶到 802 行（R9 上限 800）。不塞白名单
（棘轮只减不增），按 P2 真拆：那组 gate id 计算与门状态谓词是**纯计算**——不碰仓库、
不碰 renderer 桥、不闭包任何注入依赖，混在 600 行的 driver 编排工厂里既顶行数上限，
又让「门 id 规则」这件跨模块共享的事没有单一落点（productionRunService 也要用 isShotGate）。

- 新增 productionRunGateIdentity.ts（48 行）：shotGateId/isShotGate/sampleGateId/
  freezeGateId + 三个门状态谓词 + shouldSampleGate
- driverOps 802→767；service 的 isShotGate 直接从新模块 import，**不留转发壳**（P1）
- 顺带把 service 一处四行 import 压成一行：816→814，基线同步下调锁死

门 id 是持久且对外可见的标识（进 Run gates、进 MCP 投影、进深链），算法未改一字。
typecheck exit 0。
